+++ packages/components/__generated__/ChatRoomQuery.graphql.ts
@@ -1,5 +1,5 @@
 /**
- * @generated SignedSource<<1d3cf037c869a3294443cecd5eb5bc17>>
+ * @generated SignedSource<<acbfb67ef6de93f0442708700d4015b4>>
  * @lightSyntaxTransform
  * @nogrep
  */
@@ -48,6 +48,7 @@ export type ChatRoomQuery$data = {
             }
           | null
           | undefined
+        readonly unreadMessagesCount: number | null | undefined
         readonly ' $fragmentSpreads': FragmentRefs<'MessagesListFragment'>
       }
     | null
@@ -84,17 +85,24 @@ const node: ConcreteRequest = (function () {
       alias: null,
       args: null,
       kind: 'ScalarField',
-      name: '__typename',
+      name: 'unreadMessagesCount',
       storageKey: null,
     },
     v4 = {
       alias: null,
       args: null,
       kind: 'ScalarField',
+      name: '__typename',
+      storageKey: null,
+    },
+    v5 = {
+      alias: null,
+      args: null,
+      kind: 'ScalarField',
       name: 'name',
       storageKey: null,
     },
-    v5 = [
+    v6 = [
       {
         alias: null,
         args: null,
@@ -103,7 +111,7 @@ const node: ConcreteRequest = (function () {
         storageKey: null,
       },
     ],
-    v6 = {
+    v7 = {
       alias: null,
       args: null,
       concreteType: 'Profile',
@@ -112,8 +120,8 @@ const node: ConcreteRequest = (function () {
       plural: false,
       selections: [
         v2 /*: any*/,
-        v3 /*: any*/,
         v4 /*: any*/,
+        v5 /*: any*/,
         {
           alias: null,
           args: [
@@ -132,20 +140,20 @@ const node: ConcreteRequest = (function () {
           kind: 'LinkedField',
           name: 'image',
           plural: false,
-          selections: v5 /*: any*/,
+          selections: v6 /*: any*/,
           storageKey: 'image(height:100,width:100)',
         },
       ],
       storageKey: null,
     },
-    v7 = {
+    v8 = {
       alias: null,
       args: null,
       kind: 'ScalarField',
       name: 'totalCount',
       storageKey: null,
     },
-    v8 = [
+    v9 = [
       {
         kind: 'Literal',
         name: 'first',
@@ -168,6 +176,7 @@ const node: ConcreteRequest = (function () {
           plural: false,
           selections: [
             v2 /*: any*/,
+            v3 /*: any*/,
             {
               alias: null,
               args: null,
@@ -191,7 +200,7 @@ const node: ConcreteRequest = (function () {
                       kind: 'LinkedField',
                       name: 'node',
                       plural: false,
-                      selections: [v6 /*: any*/],
+                      selections: [v7 /*: any*/],
                       storageKey: null,
                     },
                   ],
@@ -227,6 +236,7 @@ const node: ConcreteRequest = (function () {
           plural: false,
           selections: [
             v2 /*: any*/,
+            v3 /*: any*/,
             {
               alias: null,
               args: null,
@@ -250,25 +260,25 @@ const node: ConcreteRequest = (function () {
                       kind: 'LinkedField',
                       name: 'node',
                       plural: false,
-                      selections: [v6 /*: any*/, v2 /*: any*/],
+                      selections: [v7 /*: any*/, v2 /*: any*/],
                       storageKey: null,
                     },
                   ],
                   storageKey: null,
                 },
-                v7 /*: any*/,
+                v8 /*: any*/,
               ],
               storageKey: null,
             },
             {
               alias: null,
-              args: v8 /*: any*/,
+              args: v9 /*: any*/,
               concreteType: 'MessageConnection',
               kind: 'LinkedField',
               name: 'allMessages',
               plural: false,
               selections: [
-                v7 /*: any*/,
+                v8 /*: any*/,
                 {
                   alias: null,
                   args: null,
@@ -302,7 +312,7 @@ const node: ConcreteRequest = (function () {
                           plural: false,
                           selections: [
                             v2 /*: any*/,
-                            v4 /*: any*/,
+                            v5 /*: any*/,
                             {
                               alias: null,
                               args: [
@@ -321,7 +331,7 @@ const node: ConcreteRequest = (function () {
                               kind: 'LinkedField',
                               name: 'image',
                               plural: false,
-                              selections: v5 /*: any*/,
+                              selections: v6 /*: any*/,
                               storageKey: 'image(height:32,width:32)',
                             },
                           ],
@@ -365,7 +375,7 @@ const node: ConcreteRequest = (function () {
                           name: 'verb',
                           storageKey: null,
                         },
-                        v3 /*: any*/,
+                        v4 /*: any*/,
                       ],
                       storageKey: null,
                     },
@@ -409,7 +419,7 @@ const node: ConcreteRequest = (function () {
             },
             {
               alias: null,
-              args: v8 /*: any*/,
+              args: v9 /*: any*/,
               filters: null,
               handle: 'connection',
               key: 'chatRoom_allMessages',
@@ -422,16 +432,16 @@ const node: ConcreteRequest = (function () {
       ],
     },
     params: {
-      cacheID: '9bf20b63b6591af0c171bd5d00d69793',
+      cacheID: '2898527115ec10a9725f8a0b31d7a02c',
       id: null,
       metadata: {},
       name: 'ChatRoomQuery',
       operationKind: 'query',
-      text: 'query ChatRoomQuery(\n  $roomId: ID!\n) {\n  chatRoom(id: $roomId) {\n    id\n    participants {\n      edges {\n        node {\n          profile {\n            id\n            __typename\n            name\n            image(width: 100, height: 100) {\n              url\n            }\n          }\n          id\n        }\n      }\n    }\n    ...MessagesListFragment\n  }\n}\n\nfragment MessageItemFragment on Message {\n  id\n  content\n  created\n  extraData\n  inReplyTo {\n    id\n  }\n  pk\n  profile {\n    id\n  }\n  verb\n}\n\nfragment MessagesListFragment on ChatRoom {\n  id\n  participants {\n    totalCount\n  }\n  allMessages(first: 20) {\n    totalCount\n    edges {\n      node {\n        id\n        created\n        profile {\n          id\n          name\n          image(height: 32, width: 32) {\n            url\n          }\n        }\n        ...MessageItemFragment\n        __typename\n      }\n      cursor\n    }\n    pageInfo {\n      hasNextPage\n      endCursor\n    }\n  }\n}\n',
+      text: 'query ChatRoomQuery(\n  $roomId: ID!\n) {\n  chatRoom(id: $roomId) {\n    id\n    unreadMessagesCount\n    participants {\n      edges {\n        node {\n          profile {\n            id\n            __typename\n            name\n            image(width: 100, height: 100) {\n              url\n            }\n          }\n          id\n        }\n      }\n    }\n    ...MessagesListFragment\n  }\n}\n\nfragment MessageItemFragment on Message {\n  id\n  content\n  created\n  extraData\n  inReplyTo {\n    id\n  }\n  pk\n  profile {\n    id\n  }\n  verb\n}\n\nfragment MessagesListFragment on ChatRoom {\n  id\n  participants {\n    totalCount\n  }\n  allMessages(first: 20) {\n    totalCount\n    edges {\n      node {\n        id\n        created\n        profile {\n          id\n          name\n          image(height: 32, width: 32) {\n            url\n          }\n        }\n        ...MessageItemFragment\n        __typename\n      }\n      cursor\n    }\n    pageInfo {\n      hasNextPage\n      endCursor\n    }\n  }\n}\n',
     },
   }
 })()
 
-;(node as any).hash = '05693484a73cde0be72b71dabab2833a'
+;(node as any).hash = '8ac24822b8b22e3b11eea731ab254d9e'
 
 export default node