+++ packages/components/modules/messages/CreateChatRoomList/ChatRoomListCard/index.tsx
@@ -3,22 +3,25 @@
 import { FC } from 'react'
 
 import { AvatarWithPlaceholder } from '@baseapp-frontend/design-system'
+import { joinWithSeparator } from '@baseapp-frontend/utils'
 
 import { LoadingButton } from '@mui/lab'
 import { Box, Typography } from '@mui/material'
 
+import { useChatRoom } from '../../context'
 import { MainContainer } from './styled'
 import { ChatRoomListCardProps } from './types'
 
 const ChatRoomListCard: FC<ChatRoomListCardProps> = ({
   item,
-  setIsInChatRoom,
   setIsInExistingChatRoomsView,
   currentProfile,
   commit,
   isMutationInFlight,
 }) => {
   const { id, image, name, urlPath } = item
+
+  const { setChatRoom } = useChatRoom()
 
   return (
     <MainContainer key={`profile-${id}`}>
@@ -31,7 +34,7 @@ const ChatRoomListCard: FC<ChatRoomListCardProps> = ({
       <Box sx={{ display: 'grid', gridTemplateRows: 'repeat(2, minmax(0, 1fr))' }}>
         <Typography variant="subtitle2">{name}</Typography>
         <Typography variant="caption" color="text.secondary">
-          {urlPath?.path && `@${urlPath?.path}`}
+          {joinWithSeparator(['@', urlPath?.path], { separator: '' })}
         </Typography>
       </Box>
       <LoadingButton
@@ -43,8 +46,8 @@ const ChatRoomListCard: FC<ChatRoomListCardProps> = ({
               variables: {
                 input: { profileId: currentProfile.profile.id, participants: [id] },
               },
-              onCompleted: () => {
-                setIsInChatRoom(true)
+              onCompleted: (data) => {
+                setChatRoom({ id: data?.chatRoomCreate?.room?.node?.id })
                 setIsInExistingChatRoomsView(true)
               },
             })