+++ packages/components/modules/messages/CreateChatRoomList/ChatRoomListCard/types.ts
@@ -1,13 +1,12 @@
-import { Dispatch, PropsWithChildren, SetStateAction } from 'react'
+import { Dispatch, SetStateAction } from 'react'
 
 import { Disposable, UseMutationConfig } from 'react-relay'
 
 import { CreateChatRoomMutation } from '../../../../__generated__/CreateChatRoomMutation.graphql'
 import { UseCurrentProfile } from '../../../profiles/context/CurrentProfileProvider/types'
 
-export interface ChatRoomListCardProps extends PropsWithChildren {
+export interface ChatRoomListCardProps {
   item: any
-  setIsInChatRoom: Dispatch<SetStateAction<boolean>>
   setIsInExistingChatRoomsView: Dispatch<SetStateAction<boolean>>
   isMutationInFlight: boolean
   commit: (config: UseMutationConfig<CreateChatRoomMutation>) => Disposable