+++ packages/components/modules/messages/index.ts
@@ -7,11 +7,14 @@ export type * from './MessagesList/types'
 export { default as ChatRoom } from './ChatRoom'
 export type * from './ChatRoom/types'
 
+export { default as ChatRoomsList } from './ChatRoomsList'
+export type * from './ChatRoomsList/types'
+
 export { default as SendMessage } from './SendMessage'
 export type * from './SendMessage/types'
+
 export { default as CreateChatRoomList } from './CreateChatRoomList'
 export type * from './CreateChatRoomList/types'
-
 export * from './context'
 
 export * from './graphql/mutations/SendMessage'
@@ -20,5 +23,7 @@ export * from './graphql/mutations/CreateChatRoom'
 export * from './graphql/queries/MessageItem'
 export * from './graphql/queries/MessagesList'
 export * from './graphql/queries/ChatRoomQuery'
+export * from './graphql/queries/Room'
+export * from './graphql/queries/RoomsList'
 
 export { default as useMessagesListSubscription } from './graphql/subscriptions/useMessagesListSubscription'