+++ packages/design-system/components/Searchbar/types.ts
@@ -1,13 +1,9 @@
 import { TransitionStartFunction } from 'react'
 
-import { FieldValues, UseFormReturn } from 'react-hook-form'
-
 import type { TextFieldProps } from '../inputs/TextField/types'
 
-export type SearchbarProps = Omit<TextFieldProps, 'name'> & {
+export type SearchbarProps = TextFieldProps & {
   isPending: boolean
   refetch: any
   startTransition: TransitionStartFunction
-  form: UseFormReturn<FieldValues, any, any>
-  name: string
 }