+++ packages/design-system/components/inputs/TextField/index.tsx
@@ -7,7 +7,7 @@ import { withController } from '@baseapp-frontend/utils'
 import { useTheme } from '@emotion/react'
 import { TextField as MUITextField, Theme, useMediaQuery } from '@mui/material'
 
-import { TextFieldProps } from './types'
+import { PureTextFieldProps, TextFieldProps } from './types'
 
 const TextField: FC<TextFieldProps> = ({ isResponsive = true, ...props }) => {
   const theme = useTheme() as Theme
@@ -20,4 +20,4 @@ export default withController(TextField)
 
 // exporting the TextField without the controller, it's useful when a component
 // that inherits from TextField needs to use the same controller
-export const PureTextField = TextField
+export const PureTextField = TextField as FC<PureTextFieldProps>